+++ src/config.js
@@ -1,6 +1,6 @@
 'use strict'
 
-export const VERSION = '2.2.5'
+export const VERSION = '2.2.6'
 export const INFINITY = 1 / 0
 export const MAX_SAFE_INTEGER = 9007199254740991
 export const MAX_INTEGER = 1.7976931348623157e+308