+++ src/format/index.js
@@ -3,10 +3,11 @@
 import num, {numRus} from './num'
 import intWord from './intWord'
 import fileSize from './fileSize'
+import {padDateTime, padNumber} from './pad'
 
 export {
-    num,
-    numRus,
+    num, numRus,
     fileSize,
     intWord,
+    padDateTime, padNumber,
 }