+++ src/number/index.js
@@ -1,9 +1,7 @@
 'use strict'
 
-import pad from './pad'
 import toNumber from './toNumber'
 
 export {
-    pad,
     toNumber,
 }