+++ tests/core/equals.js
@@ -35,6 +35,16 @@ describe('equals', () => {
         expect(equals(2.211, 2.21)).toBeFalsy()
     })
 
+    it('compare object', () => {
+        expect(equals({}, {})).toBeTruthy()
+        expect(equals({k: 1}, {k: 1})).toBeTruthy()
+        expect(equals({k: 1, v: []}, {k: 1, v: []})).toBeTruthy()
+        expect(equals({k: 1, v: null}, {k: 1, v: null})).toBeTruthy()
+        expect(!equals({k: 1, v: null}, {k: 1, v: undefined})).toBeTruthy()
+        expect(equals({k: 1, v: [12, 3]}, {k: 1, v: [12, 3]})).toBeTruthy()
+        expect(equals({k: 1, v: [12, 3], items: {k: 3}}, {k: 1, v: [12, 3], items: {k: 3}})).toBeTruthy()
+    })
+
     it('compare arrays', () => {
         expect(equals([1, 2, 3, 4], [1, 2, 3, 4])).toBeTruthy()
         expect(equals([], [])).toBeTruthy()
@@ -43,5 +53,25 @@ describe('equals', () => {
 
         expect(equals([1, 2, 3, []], [1, 2, 3, []])).toBeTruthy()
         expect(equals([1, 2, 3, [5, 6]], [1, 2, 3, [5, 6]])).toBeTruthy()
+    })
+
+    it('compare bool', () => {
+        expect(equals(true, true)).toBeTruthy()
+        expect(equals(false, false)).toBeTruthy()
+        expect(!equals(false, true)).toBeTruthy()
+        expect(!equals(true, false)).toBeTruthy()
+    })
+
+    it('compare functions', () => {
+        const dFn = new Date()
+        expect(equals(() => {
+        }, () => {
+        })).toBeTruthy()
+        expect(equals(dFn, dFn)).toBeTruthy()
+        expect(equals(new Date(Date.parse('1970-01-01T00:00:00')), new Date(Date.parse('1970-01-01T00:00:00')))).toBeTruthy()
+        expect(equals(new Date(123), new Date(123))).toBeTruthy()
+        expect(equals(new Function(), new Function())).toBeTruthy()
+        expect(!equals(() => {
+        }, new Function())).toBeTruthy()
     })
 })